# swagger.yaml
openapi: 3.0.3
info:
  title: QuizLab API
  description: |
    API REST complète pour la gestion de quiz avec système de cryptage automatique des IDs.
  version: 1.0.0
  contact:
    name: API Support
    email: support@quizlab.com
  license:
    name: MIT
servers:
  - url: /api/v1
    description: "Base path for API (routes mounted under /api/v1 in server.ts)"
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erreur lors de l'opération"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    HashId:
      type: string
      description: ID crypté avec HashIds (8+ caractères alphanumériques)
      example: "VAG7e3Qj"
      pattern: "^[A-Za-z0-9]{8,}$"
...
    Participation:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/HashId'
        quiz_id:
          $ref: '#/components/schemas/HashId'
        utilisateur_id:
          $ref: '#/components/schemas/HashId'
        email_participant:
          type: string
          format: email
        nom_participant:
          type: string
        statut:
          type: string
          enum: [en_cours, termine, abandonne]
          example: "en_cours"
        score:
          type: integer
          nullable: true
          example: 8
        score_max:
          type: integer
          nullable: true
          example: 10
        pourcentage:
          type: integer
          nullable: true
          example: 80
        temps_total:
          type: integer
          nullable: true
          description: Temps total en secondes
        date_debut:
          type: string
          format: date-time
        date_fin:
          type: string
          format: date-time
          nullable: true
paths:
...
  /quizzes:
    get:
      tags: [Quiz]
      summary: "📋 Liste des quiz PUBLICS (sans authentification)"
      description: |
        ## 🌍 Endpoint PUBLIC - Aucune authentification requise
        
        Retourne UNIQUEMENT les quiz **publiés** (`statut: "publie"`).
        
        **❌ Informations EXCLUES (sécurité) :**
        - `est_correcte` dans les choix de réponses (bonnes réponses cachées)
        - `invitations` (données privées du créateur)
        - `participations` (données privées)
        - `email` du créateur (vie privée)
        
        **✅ Informations INCLUSES :**
        - Titre, description du quiz
        - Questions et choix de réponses (sans les bonnes réponses)
        - Nom/prénom du créateur
        - IDs cryptés automatiquement
        
        **Cas d'usage :** Page d'accueil avec liste des quiz disponibles
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          example: 10
          description: Nombre maximum de quiz à retourner
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination - position de départ
      responses:
        '200':
          description: Liste des quiz publics (sans bonnes réponses, invitations, participations)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      quizzes:
                        type: array
                        description: Quiz publics publiés
                        items:
                          type: object
                          properties:
                            id:
                              $ref: '#/components/schemas/HashId'
                            titre:
                              type: string
                            description:
                              type: string
                            type_quiz:
                              type: string
                            statut:
                              type: string
                              example: "publie"
                            createur:
                              type: object
                              properties:
                                id:
                                  $ref: '#/components/schemas/HashId'
                                prenom:
                                  type: string
                                nom:
                                  type: string
                            questions:
                              type: array
                              items:
                                type: object
                                properties:
                                  id:
                                    $ref: '#/components/schemas/HashId'
                                  texte:
                                    type: string
                                  duree:
                                    type: integer
                                  ordre:
                                    type: integer
                                  choix_reponses:
                                    type: array
                                    description: "⚠️ SANS est_correcte (bonnes réponses cachées)"
                                    items:
                                      type: object
                                      properties:
                                        id:
                                          $ref: '#/components/schemas/HashId'
                                        texte:
                                          type: string
                                        ordre:
                                          type: integer
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      total:
                        type: integer
              example:
                success: true
                data:
                  quizzes:
                    - id: "VAG7e3Qj"
                      titre: "Quiz JavaScript"
                      statut: "publie"
                      createur:
                        id: "2m3KMGnQ"
                        prenom: "Jean"
                        nom: "Dupont"
                      questions:
                        - id: "9x360G5g"
                          texte: "Quelle est la différence entre let et var ?"
                          duree: 30
                          ordre: 1
                          choix_reponses:
                            - id: "4JGR4Q36"
                              texte: "let a une portée de bloc"
                              ordre: 1
                            - id: "6bPLmwGO"
                              texte: "var a une portée de fonction"
                              ordre: 2
                pagination:
                  page: 1
                  total: 5
...
  /participations:
    post:
      tags: [Participations]
      summary: "CAS 2: Démarrer une participation avec CODE D'INVITATION"
      description: |
        ## ✉️ Participation via invitation (quiz privé OU public)

        **Utilisé pour:**
        - Quiz avec invitations envoyées par email
        - Quiz privés (`est_public: false`) - UNIQUEMENT accessible par invitation
        - Quiz publics avec invitations personnalisées

        **Workflow complet d'invitation :**
        1. Créateur crée invitation → `POST /invitations/quizzes/{quizId}/invitations` (email envoyé automatiquement)
        2. Participant reçoit email avec `code_acces` et lien d'invitation
        3. Participant clique sur le lien ou utilise le code directement
        4. Participant valide le code → `POST /invitations/validate` (optionnel)
        5. Participant démarre participation avec cet endpoint
        6. Backend récupère automatiquement le `quiz_id` depuis l'invitation

        **⚠️ DIFFÉRENCE CLÉS:**
        - `quiz_id` N'EST PAS requis (calculé depuis l'invitation)
        - Fonctionne même si le quiz est privé
        - Code d'invitation doit être valide et non expiré
        - **Email automatique :** Un email de confirmation peut être envoyé lors de la participation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code_acces]
              properties:
                code_acces:
                  type: string
                  description: Code d'invitation reçu par email (32 caractères hexadécimaux)
                  example: "df22be74d996f008cb04ee412b5f1aa1"
                email_participant:
                  type: string
                  format: email
                  description: Email du participant (optionnel si déjà dans l'invitation)
                nom_participant:
                  type: string
                  description: Nom du participant (optionnel)
            example:
              code_acces: "df22be74d996f008cb04ee412b5f1aa1"
              email_participant: "invite@example.com"
      responses:
        '201':
          description: Participation démarrée via invitation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      participation:
                        $ref: '#/components/schemas/Participation'
              example:
                success: true
                message: "Participation démarrée avec succès"
                data:
                  participation:
                    id: "Rb3Wb19k"
                    quiz_id: "VAG7e3Qj"
                    statut: "en_cours"
        '400':
          description: Code d'invitation invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                code_invalide:
                  summary: Code invalide
                  value:
                    success: false
                    message: "Code d'invitation invalide"
                code_expire:
                  summary: Code expiré
                  value:
                    success: false
                    message: "Ce code d'invitation a expiré"
...